requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-genai>=1.0.0
playwright>=1.40.0
Pillow>=10.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
